+++ report-service/src/main/java/com/example/reportservice/controller/VisitAuthRequestController.java
@@ -2,8 +2,10 @@ package com.example.reportservice.controller;
 
 import com.example.reportservice.common.aop.AdminUser;
 import com.example.reportservice.common.constant.VisitAuthConstant;
+import com.example.reportservice.common.utils.ImageUtils;
 import com.example.reportservice.dto.visit_auth.VisitAuthRequestDetailDTO;
 import com.example.reportservice.dto.visit_auth.VisitAuthRequestResponseDTO;
+import com.example.reportservice.service.s3.S3Service;
 import com.example.reportservice.service.visit_auth.VisitAuthRequestService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 
 @RestController
@@ -22,14 +25,22 @@ import java.util.List;
 public class VisitAuthRequestController {
 
     private final VisitAuthRequestService visitAuthRequestService;
+    private final S3Service s3Service;
 
     @PostMapping(value = "{culturalEventId}",consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<Void> createVisitAuthRequest(final @RequestHeader("userId") long userId,
                                                  final @PathVariable int culturalEventId,
                                                  final @RequestPart("fileList") List<MultipartFile> fileList) {
 
+        if(!ImageUtils.validateImage(fileList)) {
+            throw new IllegalArgumentException("Invalid image file");
+        }
 
-        visitAuthRequestService.createVisitAuthRequest(userId, culturalEventId, fileList);
+        final List<String> imageUrls = fileList.stream()
+                .map(s3Service::uploadFile)
+                .collect(Collectors.toList());
+
+        visitAuthRequestService.createVisitAuthRequest(userId, culturalEventId, imageUrls);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
@@ -43,6 +54,7 @@ public class VisitAuthRequestController {
     @GetMapping("{visitAuthId}")
     @AdminUser
     public ResponseEntity<VisitAuthRequestDetailDTO> getVisitAuthRequest(final @PathVariable int visitAuthId) {
+
         return ResponseEntity.ok().body(visitAuthRequestService.getVisitAuthRequest(visitAuthId));
     }
