+++ report-service/src/main/java/com/example/reportservice/domain/adapter/visit_auth/query/WhereQuery.java
@@ -1,9 +1,9 @@
-package com.example.reportservice.repository.visit_auth.query;
+package com.example.reportservice.domain.adapter.visit_auth.query;
 
 import com.example.reportservice.common.constant.VisitAuthConstant;
 import com.querydsl.core.types.dsl.BooleanExpression;
 
-import static com.example.reportservice.entity.visit_auth.QVisitAuthRequest.visitAuthRequest;
+import static com.example.reportservice.domain.entity.visit_auth.QVisitAuthRequest.visitAuthRequest;
 
 public final class WhereQuery {
