+++ report-service/src/main/java/com/example/reportservice/domain/entity/outbox/OutBox.java
@@ -1,4 +1,4 @@
-package com.example.reportservice.entity.outbox;
+package com.example.reportservice.domain.entity.outbox;
 
 import jakarta.persistence.*;
 import lombok.*;
