+++ report-service/src/main/java/com/example/reportservice/dto/visit_auth/VisitAuthRequestDetailDTO.java
@@ -1,13 +1,12 @@
 package com.example.reportservice.dto.visit_auth;
 
-import com.example.reportservice.entity.event_report.CulturalEventDetail;
-import com.example.reportservice.entity.visit_auth.VisitAuthRequest;
+import com.example.reportservice.domain.entity.event_report.CulturalEventDetail;
+import com.example.reportservice.domain.entity.visit_auth.VisitAuthRequest;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @AllArgsConstructor
 @Builder
@@ -19,8 +18,8 @@ public class VisitAuthRequestDetailDTO {
     private LocalDateTime createdAt;
     private boolean isAuthenticated;
 
-    public static VisitAuthRequestDetailDTO of(final VisitAuthRequest visitAuthRequest, final CulturalEventDetail culturalEventDetail) {
-        return VisitAuthRequestDetailDTO.builder()
+    public VisitAuthRequestDetailDTO(final VisitAuthRequest visitAuthRequest, final CulturalEventDetail culturalEventDetail) {
+        VisitAuthRequestDetailDTO.builder()
                 .id(visitAuthRequest.getId())
                 .createdAt(visitAuthRequest.getCreatedAt())
                 .isAuthenticated(visitAuthRequest.isAuthenticated())
